LayoutTests/imported/w3c/web-platform-tests/html/browsers/origin/origin-keyed-agent-clusters/getter-special-cases/cross-origin-isolated.sub.https-expected.txt



FAIL self: originAgentCluster must equal true assert_equals: expected (boolean) true but got (undefined) undefined
FAIL child: originAgentCluster must equal true assert_equals: expected (boolean) true but got (undefined) undefined
